move pkg/startup to core/startup -- it does have important and unfixable dependencies on core
